docs(realtime): fix bigint-cast consistency nit, sync deferred-findings doc
Found by adversarial review: the match-count write-back's last_matched_at
bind was missing the same ::bigint cast every other scalar bind in this
UPDATE already carries -- functionally harmless, but inconsistent. Also
updates docs/realtime-firehose.md's deferred-findings section, which still
listed #5022/#5023/#5024 as open after this commit fixes all three.

+++ workers/data-api.mjs
@@ -2529,7 +2529,7 @@ async function handleAlertTriggersMatchedWriteback(request, env) {
     const updated = await sql.unsafe(
       `UPDATE chain_alert_triggers
        SET match_count = match_count + 1,
-           last_matched_at = $1
+           last_matched_at = $1::bigint
        WHERE id IN (${placeholders})
        RETURNING id`,
       [now, ...ids],
